backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
stripe==8.9.0
python-dotenv==1.0.1
pydantic==2.11.3
beanie>=1.23.6
motor>=3.3.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.1
python-multipart==0.0.9
pytest==8.2.1
httpx==0.27.0
